I _believe_ that's all the relevant updates to the Firepad library itself. Left several things from the old version of the library (from firepad's 'atom' branch, which is a little old...) in comments in this version, just in case.
