+++ lib/lib.bak.js
@@ -2036,9 +2036,9 @@ firepad.Client = (function () {
   // hasn't yet received our newest operation, an insertion of 5 characters at
   // the beginning of the document, the correct position of the other user's
   // cursor in our current document is 8.
-  Client.prototype.transformCursor = function (cursor) {
-    return this.state.transformCursor(cursor);
-  };
+ // Client.prototype.transformCursor = function (cursor) {
+ //   return this.state.transformCursor(cursor);
+ // };
 
   Client.prototype.serverRetry = function() {
     this.setState(this.state.serverRetry(this));
@@ -2129,10 +2129,11 @@ firepad.Client = (function () {
     client.sendOperation(this.outstanding);
     return this;
   };
-
-  AwaitingConfirm.prototype.transformCursor = function (cursor) {
-    return cursor.transform(this.outstanding);
-  };
+  
+  //oldlibs
+  // AwaitingConfirm.prototype.transformCursor = function (cursor) {
+  //   return cursor.transform(this.outstanding);
+  // };
 
   // In the 'AwaitingWithBuffer' state, the client is waiting for an operation
   // to be acknowledged by the server while buffering the edits the user makes
@@ -2187,10 +2188,10 @@ firepad.Client = (function () {
     client.sendOperation(this.buffer);
     return new AwaitingConfirm(this.buffer);
   };
-
-  AwaitingWithBuffer.prototype.transformCursor = function (cursor) {
-    return cursor.transform(this.outstanding).transform(this.buffer);
-  };
+  // oldlibs
+  // AwaitingWithBuffer.prototype.transformCursor = function (cursor) {
+  //   return cursor.transform(this.outstanding).transform(this.buffer);
+  // };
 
   return Client;
 
@@ -4893,14 +4894,14 @@ firepad.Firepad = (function(global) {
     }
 
     if (!this.atom_) {
-      var editorWrapper = this.codeMirror_ ? this.codeMirror_.getWrapperElement() : this.ace_.container;
-      this.firepadWrapper_ = utils.elt("div", null, { 'class': 'firepad' });
-      editorWrapper.parentNode.replaceChild(this.firepadWrapper_, editorWrapper);
-      this.firepadWrapper_.appendChild(editorWrapper);
+    var editorWrapper = this.codeMirror_ ? this.codeMirror_.getWrapperElement() : this.ace_.container;
+    this.firepadWrapper_ = utils.elt("div", null, { 'class': 'firepad' });
+    editorWrapper.parentNode.replaceChild(this.firepadWrapper_, editorWrapper);
+    this.firepadWrapper_.appendChild(editorWrapper);
 
-      // Don't allow drag/drop because it causes issues.  See https://github.com/firebase/firepad/issues/36
-      utils.on(editorWrapper, 'dragstart', utils.stopEvent);
-    }
+    // Don't allow drag/drop because it causes issues.  See https://github.com/firebase/firepad/issues/36
+    utils.on(editorWrapper, 'dragstart', utils.stopEvent);
+  }
 
     // Provide an easy way to get the firepad instance associated with this CodeMirror instance.
     this.editor_.firepad = this;
@@ -5519,6 +5520,7 @@ firepad.Firepad = (function(global) {
     this.firepadWrapper_.insertBefore(this.toolbar.element(), this.firepadWrapper_.firstChild);
   };
 
+/* oldlib
   Firepad.prototype.registerBuiltinEntities_ = function() {
     var attrs = ['src', 'alt', 'width', 'height', 'style', 'class'];
     this.registerEntity('img', {
@@ -5547,6 +5549,7 @@ firepad.Firepad = (function(global) {
       }
     });
   };
+*/
 
   Firepad.prototype.addPoweredByLogo_ = function() {
     var poweredBy = utils.elt('a', null, { 'class': 'powered-by-firepad'} );
